requirements.txt
PySide6
APScheduler
pyyaml
selenium
webdriver-manager
requests
pandas
openpyxl
ddddocr
Pillow
